added global middleware condensed exsiting middleware deleted unwanted files and excess code created a job id middleware function
