+++ jobs/jobs-model.js
@@ -31,7 +31,5 @@ module.exports = {
   //get all jobs by userId
   getJobsByUserId(clientId) {
     return knex("jobs").where({ clientId });
-
   },
-
 };
