+++ users/users-router.js
@@ -40,12 +40,10 @@ usersRouter.put("/rating/:id", (req, res) => {
             res.status(201).json(user);
           })
           .catch((err) => {
-            res
-              .status(500)
-              .json({
-                message: "error in updating the user",
-                error: err.message,
-              });
+            res.status(500).json({
+              message: "error in updating the user",
+              error: err.message,
+            });
           });
       } else {
         // adds rating if it is the first one for the user
@@ -55,12 +53,10 @@ usersRouter.put("/rating/:id", (req, res) => {
           })
           .catch((err) => {
             console.log(err);
-            res
-              .status(500)
-              .json({
-                message: "error updating the user rating",
-                error: err.message,
-              });
+            res.status(500).json({
+              message: "error updating the user rating",
+              error: err.message,
+            });
           });
       }
     })
